Don't assume PHPARKITECT_COMPOSER_INSTALL is not defined yet
By default, the classmap is not loaded, so it's not possible to
use functions like `is_a` to infer inheritance of classes under test.
To solve this, the autoloader must be included in the config file,
but then it will try to define this variable again, which
will cause it to throw an error.